Move rule definitions into `_rules()`
Currently our `_schema()` object defines the schema as a set of rules. Since many of these will be common to both presroc and sroc, we want to be able to reuse them to avoid defining them twice. We therefore move the rule definitions into `_rules()`, which returns an object containing the rules so that once we have separate schema for presroc and sroc, we can pick and choose which rules to include in each and we only need to define each one once.