Move `ruleset` default from translator to controller
Because the presroc charge translator is only called when the ruleset is `presroc`, we don't need to set a default value. In anticipation of creating separate controllers for v2 (which defaults to `presroc`) and v3 (which sets no default as this will be a required field), we set the default in the controller instead. We also update our fixtures accordingly.