Create `CalculatePresrocChargeTranslator` (#592)
* Update Calculate Charge Translator to support SRoC

https://eaflood.atlassian.net/jira/software/projects/CMEA/boards/907?selectedIssue=CMEA-194

We validate incoming charge calculation data using CalculateChargeTranslator, which currently only supports pre-SRoC requests. We rename it CalculatePresrocChargeTranslator and create a separate CalculateSrocChargeTranslator to handle incoming SRoC validation.

As part of the change, we have identified several fields where the validation rules are the same. We will also be creating a separate CalculatePresrocChargeV2Translator at a later date to take into account some differences between v2 and v3 presroc validation (eg. we will no longer be capitalising certain strings such as season, in order to keep v3 sroc and presroc translation consistent). We therefore recognise the need to refactor these common rules, which will be done in a future PR.

* Move rule definitions into `_rules()`

Currently our `_schema()` object defines the schema as a set of rules. Since many of these will be common to both presroc and sroc, we want to be able to reuse them to avoid defining them twice. We therefore move the rule definitions into `_rules()`, which returns an object containing the rules so that once we have separate schema for presroc and sroc, we can pick and choose which rules to include in each and we only need to define each one once.

* Create `CalculatePresrocChargeTranslator`

We will be setting up separate translators for presroc and sroc charge requests, which will extend a `CalculateChargeTranslator` so they can pool common functionality. The first step is to move the existing `CalculateChargeTranslator` to `CalculatePresrocChargeTranslator`

* Move `ruleset` default from translator to controller

Because the presroc charge translator is only called when the ruleset is `presroc`, we don't need to set a default value. In anticipation of creating separate controllers for v2 (which defaults to `presroc`) and v3 (which sets no default as this will be a required field), we set the default in the controller instead. We also update our fixtures accordingly.

* Write validation rules

* Add ruleset validation to presroc validator

Although `ruleset` should be set to `presroc` when calling the presroc validator, we update the validation to check for this just to be certain (in the same way that the sroc validator checks the ruleset is `sroc`).