Change `CalculateChargeService` to use ruleset
We want `CalculateChargeService` to use different translators and presenters depending on whether this is a presroc or sroc request. The first step towards this is to add in a switch statement to assign the translators and presenters for a presroc request (ie. the only kind of request we currently support).

At this stage we haven't yet started renaming the translators and presenters (ie. `RulesServiceTranslator` will become `RulesServicePresrocTranslator`) so for the time being we simply rename when we destructure them.