Amend `compensationChargePercent` in translator
After creating the translator, we had second thoughts about the value that `compensationChargePercent` returns. We had it confirmed that it should return the percentage number, rather than a decimal value, eg. if we receive `50%` then we should return `50` rather than `0.5`. We had taken the latter approach and therefore we amend the translator accordingly.