Calculate standalone SRoC charge (#595)
https://eaflood.atlassian.net/jira/software/projects/CMEA/boards/907?selectedIssue=CMEA-193

This PR handles the bulk of the work needed for a working /v3/{regimeSlug}/calculate-charge route.

In this PR we:

create the translators and presenters needed for sroc requests;
amend CalculateChargeService to accept both presroc and sroc requests;
connect it to the v3 route;
add a guard to the v2 route to block requests with a ruleset;
remove the now-unneeded ValidateChargeService.

* Change `CalculateChargeService` to use ruleset

We want `CalculateChargeService` to use different translators and presenters depending on whether this is a presroc or sroc request. The first step towards this is to add in a switch statement to assign the translators and presenters for a presroc request (ie. the only kind of request we currently support).

At this stage we haven't yet started renaming the translators and presenters (ie. `RulesServiceTranslator` will become `RulesServicePresrocTranslator`) so for the time being we simply rename when we destructure them.

* We noticed that `RequestRulesServiceChargeService` was imported as `RequestRulesServiceCharge`. We rename it to avoid inconsistency across our services.

* Rename presroc translators and presenters

We add `presroc` to the name of the existing presenters and translators to differentiate them from the sroc presenters and translators we'll be creating.

* Rename calc. charge translators for consistency

When we created separate presroc/sroc calculate charge translators, we named them `CalculatePresrocChargeTranslator` and `CalculateSrocChargeTranslator`. However when creating our additional translators and presenters for the calculate chage update, we realise that the regime name should come last, eg. `CalculateChargePresrocTranslator`.

* Create `RulesServiceSrocPresenter`

* Create `RulesServiceSrocTranslator`

* Amend `compensationChargePercent` in translator

After creating the translator, we had second thoughts about the value that `compensationChargePercent` returns. We had it confirmed that it should return the percentage number, rather than a decimal value, eg. if we receive `50%` then we should return `50` rather than `0.5`. We had taken the latter approach and therefore we amend the translator accordingly.

* Create `CalculateChargeSrocPresenter`

* Correct translator class names

We noticed that `CalculateChargeSrocTranslator` and `RulesServiceSrocTranslator` had `Presroc` left in their class names.

* Amend calculate charge translator financial year

We created `CalculateChargeSrocTranslator` to store the financial year as `financialYear`. However `CalculateChargeService` expects this to be `chargeFinancialYear`, so we amend the translator accordingly.

* Add sroc to `CalculateChargeService`

We update `CalculateChargeService` to use the sroc presenters and translators when `ruleset` is `sroc`, and add a unit test to check this.

* Add v2 guard service and update unit tests

We add a v2 guard service to thrown an error if a request contains `ruleset` (as this would clearly be a v3 request). We add this to the unit tests, and update them in line with our current "stub services in controllers" approach.

* Remove `.only`

* Update v3 route to use `CalculateChargeController`

Now that `CalculateChargeController` supports sroc, we update the v3 route to use it

* Delete `ValidateChargeService`

Now that the v3 Calculate Charge route uses `CalculateChargeService`, we can delete `ValidateChargeService` as we no longer need it.

* Rename sroc `volume` to `authorisedVolume`

We have agreed that the sroc `volume` parameter is better named `authorisedVolume` to differentiate it from `actualVolume`.