Fix false positive unit tests
While working on the unit tests we realised that some of our "throws an error" test weren't working as intended. The tests were passing as a validation error was being thrown; however the errors were not being caused by the key being tested!

On investigation we realised that we were incorrectly setting up our `invalidPayload` for testing:
```
const invalidPayload = payload
```
Simply assigning it in this way makes `invalidPayload` a reference to the original object, so deleting a key from `invalidPayload` or changing its value to an invalid one also affects `payload`. We resolve this by ensuring we create a new object from `payload`:
```
const invalidPayload = { ...payload }
```

Doing this highlights several broken tests, so we fix the tests and the validation rules accordingly.