Update `supportedSourceName` validation
The Rules Service validates the supported source name regardless of whether or not the supported source flag is set. To resolve this, a new supported source of `Not Applicable` is added. We therefore change the validation of `supportedSourceName` so that when `supportedSource` is `false`, we must not be given a value for `supportedSourceName`, and we set its value to `Not Applicable`.