Rename test file for easier review
We found that renaming the test file and making changes to it caused the pull request diff to see it as an entirely new file without any changes highlighted. This isn't really desirably for reviewing! We therefore rename the file back to make the changes more apparent, and we will name it correctly in a separate PR.