Add additional conditional test cases
Some fields are conditional based on others. For example, if `compensationCharge` is `true` then `regionalChargingArea` and `waterUndertaker` must be present. These fields have validation and are marked as required if `compensationCharge` is `true`, but they are not explicitly marked as optional. We therefore add unit tests to explictly confirm that they do not fail validation if `compensationCharge` is `false` but the fields are not present.