Update sroc charge validation (#602)
https://eaflood.atlassian.net/jira/software/projects/CMEA/boards/907?selectedIssue=CMEA-194

We have decided to perform additional validation on v3 charge requests to help reduce errors coming back from the Rules Service:
- Certain combinations of boolean parameters are invalid, so we reject them.
- The Rules Service expects some text fields to have specific capitalisation, so we apply this before passing on the request.
- supportedSourceName is only required if supportedSource is set to true. However, the Rules Service validates that supportedSourceName is a valid string, regardless of whether supportedSource is true or not. We have agreed that the Rules Service will accept Not Applicable as a valid source, and we therefore use this value for supportedSourceName if supportedSource is false. We also return an error if supportedSource is false but supportedSourceName is populated.

While we update the validation, we make some additional changes:
- We fix an error with authorisedVolume where we set a minimum value of 0, whereas the AC actually requires it to be greater than 0.
- Some tests were passing for the wrong reason; a validation error was being thrown, but not for the parameter being tested. We therefore fix this.
- Some items are only required in certain circumstances, eg. another parameter is true. If the parameter was false then we were not doing any validation on the optional item. We change the order of validation so they are always validated, regardless of whether they are needed.
- Finally, we change the earliest periodStart date from 01-APR-2022 to 01-APR-2021. The 2022/23 financial year is the first year that SRoC requests will be valid. However, we believe that WRLS will need be able to create transactions for the 2021/22 financial year as they start developing for the v3 service. We therefore amend periodStart to allow the earlier start date. We do intend to revert this prior to going into production.

One last note: we previously renamed the translators from eg. calculate_charge_sroc.translator.js to calculate_charge_sroc.translator.js, but forgot to rename the test files to match. We initially fixed this, but found that the combination of renaming calculate_sroc_charge.translator.test.js and making changes to it led the diff view to see it as an entirely new file. We therefore revert the naming to calculate_sroc_charge.translator.test.js (leaving the presroc test file with the correct name) to make reviewing changes easier, and we will correct this in another PR.


* Rename translator test files

In a previous PR, we renamed the translator files to put the ruleset at the end eg. from `calculate_presroc_charge.translator.js` to `calculate_charge_presroc.translator.js`. However we forgot to do this to the test files! We correct this here.

* Fix false positive unit tests

While working on the unit tests we realised that some of our "throws an error" test weren't working as intended. The tests were passing as a validation error was being thrown; however the errors were not being caused by the key being tested!

On investigation we realised that we were incorrectly setting up our `invalidPayload` for testing:
```
const invalidPayload = payload
```
Simply assigning it in this way makes `invalidPayload` a reference to the original object, so deleting a key from `invalidPayload` or changing its value to an invalid one also affects `payload`. We resolve this by ensuring we create a new object from `payload`:
```
const invalidPayload = { ...payload }
```

Doing this highlights several broken tests, so we fix the tests and the validation rules accordingly.

* Fix additional test case

* Change validation order for optional rules

We move the validation out of the `.when()` block for optional rules so that validation will always be performed. So for example, if `actualVolume` is provided as a string then a validation error will always be returned, regardless of whether `twoPartTariff` is `true` or `false`.

* Improve string validation

`loss`, `regionalChargingArea` and `supportedSourceName` must all have the exact capitalisation expected by the rules service. For each of these, we add the list of valid strings to the validation and do a case-insensitive match against them -- if the string is found then it returns the correctly-cased string.

* Update `supportedSourceName` validation

The Rules Service validates the supported source name regardless of whether or not the supported source flag is set. To resolve this, a new supported source of `Not Applicable` is added. We therefore change the validation of `supportedSourceName` so that when `supportedSource` is `false`, we must not be given a value for `supportedSourceName`, and we set its value to `Not Applicable`.

* Change `periodStart` date

The 2022/23 financial year is the first year that SRoC requests will be valid. However, we believe that WRLS will need be able to create transactions for the 2021/22 financial year as they start developing for the v3 service. We are therefore amending the `periodStart` validation to allow a start date of `01-APR-2021`. We do intend to revert this prior to going into production.

* Rename test file for easier review

We found that renaming the test file and making changes to it caused the pull request diff to see it as an entirely new file without any changes highlighted. This isn't really desirably for reviewing! We therefore rename the file back to make the changes more apparent, and we will name it correctly in a separate PR.

* Add additional conditional test cases

Some fields are conditional based on others. For example, if `compensationCharge` is `true` then `regionalChargingArea` and `waterUndertaker` must be present. These fields have validation and are marked as required if `compensationCharge` is `true`, but they are not explicitly marked as optional. We therefore add unit tests to explictly confirm that they do not fail validation if `compensationCharge` is `false` but the fields are not present.