Amend sroc charge validation tests (#604)
* Amend sroc charge validation tests

In a previous PR we renamed the sroc charge validation test file and made changes to it. However this led to the pull request diff seeing this as an entirely new file. We therefore reverted the rename to make it easier to see what had changed, and we rename the file in this branch instead. We also correct an issue with incorrectly-nested blocks of tests.

* Rename test file

* Fix nesting

We realised that we had incorrectly nested our tests, so the `when the data is not valid` block was nested inside the `when the date is valid` block.