Update db for sroc transactions and charges https://eaflood.atlassian.net/jira/software/projects/CMEA/boards/907?selectedIssue=CMEA-202 https://eaflood.atlassian.net/jira/software/projects/CMEA/boards/907?selectedIssue=CMEA-193
Now that we have set up validation for sroc charges and transactions, the next step is to update the `transactions` table in the db with columns for sroc-only data items.

After doing this, we update the transaction and charge translators to use these fields, and align the remaining fields which are common to both sroc and presroc with the columns currently in use for presroc.