Change how default ruleset is set
We updated the controller to set the default ruleset of `presroc`. However as pointed out prior to review, the presroc transaction translator (currently the only one in use) sets the default ruleset anyway. We therefore change the controller back to not setting a default ruleset. We also update the translator to only allow `presroc` as a ruleset.