Change test helper default ruleset to `presroc`
Our bill run test helpers, along with `TestBillRunController`, default `ruleset` to `sroc`. However, this causes any tests which create a bill run and then create a transaction on it to fail, as we are now pulling the ruleset from the bill run but `CreateTransactionService` does not yet support sroc transactions. Rather than change every unit test that uses one of the test helpers to explicitly specify a presroc bill run, we instead change the default from `sroc` to `presroc`, allowing the tests to pass. We will revert this in a future PR once we have `CreateTransactionService` working for sroc.