Update `CreateTransactionService` to accept `ruleset` (#609)
https://eaflood.atlassian.net/jira/software/projects/CMEA/boards/907?selectedIssue=CMEA-202

Ahead of our v3 route, we update CreateTransactionService to read ruleset from the bill run that the transaction is being created on. At present, we only support presroc bill runs; sroc will be added in a future PR.

Since the v2 route optionally accepts ruleset in the payload, we update the presroc validation to only allow presroc, and to default to presroc when not given. This means that the output of the translator contains the ruleset which is needed by CalculateChargeService when we pass the validated data on to it.

Note that all of this required a change to our test helpers; every test which uses a bill run helper to create a bill run followed by CreateTransactionService would fail as they default to creating an sroc bill run, but we don't yet support creating sroc transactions. Rather than changing every single test where this occurs, then changing it back once we are able to create sroc transactions, we simply change the default bill run the helpers create to be a presroc one. This gives us only 3 small changes to revert later: the default ruleset inBillRunHelper, NewBillRunHelper, and TestBillRunController.


* `CreateTransactionService` accepts`ruleset`

We add a `switch` statement to `CreateTransactionService` to select the appropriate transaction translator. At present this only includes `presroc`; we will expand it to include `sroc` in a future PR.

* Set default `ruleset` in controller

For now we don't expect create transaction requests to contain a `ruleset`, so we set one in the controller.

* Change how default ruleset is set

We updated the controller to set the default ruleset of `presroc`. However as pointed out prior to review, the presroc transaction translator (currently the only one in use) sets the default ruleset anyway. We therefore change the controller back to not setting a default ruleset. We also update the translator to only allow `presroc` as a ruleset.

* Update service to take ruleset from bill run

We update `CreateTransactionService` to read `ruleset` from the passed-in bill run instead of the payload.

* Change test helper default ruleset to `presroc`

Our bill run test helpers, along with `TestBillRunController`, default `ruleset` to `sroc`. However, this causes any tests which create a bill run and then create a transaction on it to fail, as we are now pulling the ruleset from the bill run but `CreateTransactionService` does not yet support sroc transactions. Rather than change every unit test that uses one of the test helpers to explicitly specify a presroc bill run, we instead change the default from `sroc` to `presroc`, allowing the tests to pass. We will revert this in a future PR once we have `CreateTransactionService` working for sroc.

* Document `CreateTransactionService`

We noticed that `CreateTransactionService` was not documented in the code.