Remove unneeded translation in presroc translator
While updating `RulesServiceSrocTranslator`, we realised that we didn't need to include anything in `_translations()`, as everything returned by the translator was being set in `constructor()`. Following this, we checked `RulesServicePresrocTranslator` and noted that we had `chargeValue` in `_translations()`, but this was then being overwritten in `constructor()`. We therefore tidy up this unneeded translation.