Align sroc translator and presenter fields (#611)
* Align sroc translator and presenter fields
https://eaflood.atlassian.net/jira/software/projects/CMEA/boards/907?selectedIssue=CMEA-202

While working on Create Transaction, we realised that the various translators and presenters used in the create transaction flow aren't quite aligned in terms of their input and outputs. This PR seeks to align them by ensuring that leftover placeholder translation/presentation is replaced with the actual db fields to be used.

We found that none of the _translations() in RulesServiceSrocTranslator were needed, as all the data items returned by the translator are set in constructor(). We therefore remove them so that _translations() returns an empty object. Following this, we checked RulesServicePresrocTranslator and spotted the chargeValue was similarly unneeded, as we set it in constructor(). This too was removed from _translations().



* Update sroc presenters and translators

We update the sroc presenters and translators used in the Create Transaction flow to use db field names.

* Remove unneeded translation in presroc translator

While updating `RulesServiceSrocTranslator`, we realised that we didn't need to include anything in `_translations()`, as everything returned by the translator was being set in `constructor()`. Following this, we checked `RulesServicePresrocTranslator` and noted that we had `chargeValue` in `_translations()`, but this was then being overwritten in `constructor()`. We therefore tidy up this unneeded translation.