Fix `RegimeHelper`
We found that the test we wrote to check that `regime` was passed to `CreateTransactionService` was failing when it should be passing. On closer inspection we realised that the instance of `regime` we had contained a `pre_sroc_cutoff_date` field alongside the usual `preSrocCutoffDate`, causing the comparison assertion to fail. Looking at `RegimeHelper` we found that the insert query was inserting `pre_sroc_cutoff_date`, which was being converted  to `preSrocCutoffDate` when the insertion took place, but the object returned by Objection contained both fields. Changing it to `preSrocCutoffDate` resolves this.