Use correct ruleset in `CreateTransactionService`
While doing some manual testing of the endpoint we realised that although `CreateTransactionService` was using the bill run ruleset to determine the translator to use, it was not adding the ruleset to the request payload. Since we expect the ruleset to be present in the payload (so we can validate that the payload and translator match), we update `CreateTransactionService` to add in the ruleset. We also update our test fixtures to remove `ruleset` to avoid confusion.