Create Transaction v3 route (#610)
https://eaflood.atlassian.net/jira/software/projects/CMEA/boards/907?selectedIssue=CMEA-202

We create a v3 route for creating a transaction by simply copying the v2 route to v3. We add a guard to the v2 route to reject any attempts to create a transaction on a non-presroc bill run.

While doing this, we refactored the Create Transaction controller's tests to follow our current pattern of stubbing the services and testing they were called with the correct parameters. This highlighted a small error in RegimeHelper, which we have resolved.


* Create v2 route with guard

We create a v2 route along with `CreateTransactionV2GuardService` to reject the request if the bill run's ruleset is anything other than `sroc`. We also update the controller unit tests to use the new test helpers.

* Add sroc support to `CreateTransactionService`

We update the switch statement in `CreateTransactionService` to accept sroc and return `TransactionSrocTranslator`.

* Refactor unit tests

Instead of stubbing `RequestRulesServiceCharge` to return the presroc fixture, then restoring and re-stubbing it for the sroc test, we use Sinon matchers to determine the value of `ruleset` passed to it and return the appropriate fixture.

* Fix `RegimeHelper`

We found that the test we wrote to check that `regime` was passed to `CreateTransactionService` was failing when it should be passing. On closer inspection we realised that the instance of `regime` we had contained a `pre_sroc_cutoff_date` field alongside the usual `preSrocCutoffDate`, causing the comparison assertion to fail. Looking at `RegimeHelper` we found that the insert query was inserting `pre_sroc_cutoff_date`, which was being converted  to `preSrocCutoffDate` when the insertion took place, but the object returned by Objection contained both fields. Changing it to `preSrocCutoffDate` resolves this.

* Refactor v2 route tests

We update the v2 route tests to follow our current pattern of stubbing the services called by the controller.

* Add unit tests for v3 route

* Use correct ruleset in `CreateTransactionService`

While doing some manual testing of the endpoint we realised that although `CreateTransactionService` was using the bill run ruleset to determine the translator to use, it was not adding the ruleset to the request payload. Since we expect the ruleset to be present in the payload (so we can validate that the payload and translator match), we update `CreateTransactionService` to add in the ruleset. We also update our test fixtures to remove `ruleset` to avoid confusion.