Handle undefined `compensationChargePercentage` (#617)
We observed during testing that the Rules Service only returns a value for compensationChargePercentage if the charge request was for a compensation charge. If it wasn't a compensation charge, compensationChargePercentage is not included in the response, causing problems as we have set this as a required field.

We fix this by changing it from being a required field to one that defaults to null, and updating the translator to leave the percentage value as null in this situation instead of trying to convert it to a number.