Add `$deminimisInvoices()` to `BillRunModel`
We use the `deminimis` modifier on `InvoiceModel` in the context of getting all the deminimis invoices of a bill run when it is being generated. The deminimis value is determined by the bill run's ruleset, but updating the modifier to make use of the parent bill run's ruleset is not a straightforward task. Since we only ever use the modifier to get all of a bill run's deminimis invoices, it makes more sense to refactor it as a bill run instance method to return all the invoices. We therefore add this to `BillRunModel`.