Update `InvoiceModel.$deminimisInvoice()` method
The last step to updating our deminimis functionality to support both sroc and presroc is to amend the `InvoiceModel.$deminimisInvoice()` method to accept a deminimis limit, which it uses in place of the previous `500` presroc limit. We then update `DeleteLicenceService` (the only place we use this method) to pass in the deminimis limit of the bill run that the licence is being deleted from.