Update deminimis to support SRoC (#619)
https://eaflood.atlassian.net/jira/software/projects/CMEA/boards/907?selectedIssue=CMEA-211

SRoC invoices will have a different deminimis limit to pre-SRoC invoices. We therefore need to be able to support a different deminimis limit where required.

We initially wanted to update the deminimis modifier on InvoiceModel to take the ruleset from the parent bill run and apply the appropriate value. However this proved difficult to implement, due to the way Objection translated the query to SQL; we ran into problems when trying to apply a patch to the selected invoices.

We therefore took a fresh look at how we were using this modifier. After checking the code we saw that we only used it once, in GenerateBillRunService. And since we had the bill run there, we realised it made more sense to add a function to BillRunModel to return all the deminimis invoices on it, rather than taking the bill run and executing a relatedQuery which then had to join back to the parent bill run, etc. (note that strictly speaking, this method returns a query which can be run to retrieve the deminimis invoices, rather than the invoices themselves).

While doing this, we implemented a $deminimisLimit() method on BillRunModel which takes the bill run's ruleset and returns the appropriate deminimis limit value (which we store in StaticLookupLib).

Finally, we update the InvoiceModel.$deminimisInvoice() method to accept the value of the deminimis limit.

With all these in place, we are able to update our deminimis functionality to work for both presroc and sroc bill runs, which different deminimis limits applying to each.


* Add `$deminimisValue()` to `BillRunModel`

We add a method to the bill run model that returns the bill run's deminimis value, which will help us return all deminimis invoices on a bill run.

* Add `$deminimisInvoices()` to `BillRunModel`

We use the `deminimis` modifier on `InvoiceModel` in the context of getting all the deminimis invoices of a bill run when it is being generated. The deminimis value is determined by the bill run's ruleset, but updating the modifier to make use of the parent bill run's ruleset is not a straightforward task. Since we only ever use the modifier to get all of a bill run's deminimis invoices, it makes more sense to refactor it as a bill run instance method to return all the invoices. We therefore add this to `BillRunModel`.

* Unit test for `$deminimisInvoices()`

* Correct deminimis values

We accidentally put the wrong values for presroc and sroc deminimis.

* Replace `deminimis` modifier with method

We replace the `InvoiceModel.deminimis` modifier in `GenerateBillRunService` with the `BillRunModel.$deminimisInvoices()` method.

* Remove `deminimis` modifier

Now that we have replaced the use of the deminimis modifier on the invoice model with a function on the bill run model, we can remove it.

* Update `InvoiceModel.$deminimisInvoice()` method

The last step to updating our deminimis functionality to support both sroc and presroc is to amend the `InvoiceModel.$deminimisInvoice()` method to accept a deminimis limit, which it uses in place of the previous `500` presroc limit. We then update `DeleteLicenceService` (the only place we use this method) to pass in the deminimis limit of the bill run that the licence is being deleted from.

* Rename `deminimisValue` to `deminimisLimit`

We thought that "deminimis limit" probably reflects the intent of this better than "deminimis value".

* Fix "original invoice"

We spotted that we weren't excluding rebill/cancel invoices from the deminimis selection in `BillRunModel`. We fix this and update the unit tests to confirm this.

* Amend unit test

We add comments to the `$deminimisInvoices()` unit tests to clarify the purpose of each added invoice. We also update the assertions to specifically check that the correct deminimis invoices are returned.