Make v3 endpoint same as v2
When we revisited our versioning docs we realised that in cases such as this, where this is no change in behaviour between v2 and v3 (other than v3 including sroc support), we would simply point the v3 route to the existing controller. We therefore strip out everything needed to support separate v2 and v3 endpoints (the guard service etc.).