Consolidate v2 and v3 tests
Now that both routes use the same endpoint, we realised that we have the exact same tests for each version. We therefore put a single set of tests inside a `for` loop to run them for each versioned route. Note that we use `for...in` rather than `.forEach()` as the latter doesn't support asyncronous operations.