Generate Bill Run Summary v3 endpoint (#622)
* Create v2 guard service

We create `GenerateBillRunV2GuardService` to ensure that v2 requests are rejected if they are for an sroc bill run.

* Create v3 route and endpoint

We create our new v3 route along with its controller, and update the existing v2 controller to include `GenerateBillRunV2GuardService`. We also refactor the existing unit tests to reflect our current approach of stubbing all services. This allows us to further refactor the controller tests to remove Nock and fixtures.

* Add notifier check to unit tests

* Make v3 endpoint same as v2

When we revisited our versioning docs we realised that in cases such as this, where this is no change in behaviour between v2 and v3 (other than v3 including sroc support), we would simply point the v3 route to the existing controller. We therefore strip out everything needed to support separate v2 and v3 endpoints (the guard service etc.).

* Consolidate v2 and v3 tests

Now that both routes use the same endpoint, we realised that we have the exact same tests for each version. We therefore put a single set of tests inside a `for` loop to run them for each versioned route. Note that we use `for...in` rather than `.forEach()` as the latter doesn't support asyncronous operations.