Make `minimumChargeInvoice` conditional
The key difference between presroc and sroc for View Bill Run is that `minimumChargeInvoice` is only included in the output when viewing a presroc bill run. We do this by updating `ViewBillRunService` to pass `ruleset ` to `ViewBillRunPresenter`. In turn, this passes it to `ViewBillRunInvoicePresenter` (note that presenters can only accept a single `data` parameter so we need to add this to the data we pass in). Finally, this presenter excludes it from the output if the ruleset isn't `presroc`.