View SRoC Bill Run and v3 route (#624)
https://eaflood.atlassian.net/jira/software/projects/CMEA/boards/907?selectedIssue=CMEA-215

The key difference between presroc and sroc for View Bill Run is that minimumChargeInvoice is only included in the output when viewing a presroc bill run. We do this by updating ViewBillRunService to pass ruleset to ViewBillRunPresenter. In turn, this passes it to ViewBillRunInvoicePresenter (note that presenters can only accept a single data parameter so we need to add this to the data object we pass in). Finally, the invoice presenter only includes minimumChargeInvoice in the output if the ruleset is presroc.

Once this is done, we add a v3 route and unit test. Note that we do not see the changes made here to be a breaking change, so the v2 and v3 routes both use the same controller and therefore have the same functionality.


* Make `minimumChargeInvoice` conditional

The key difference between presroc and sroc for View Bill Run is that `minimumChargeInvoice` is only included in the output when viewing a presroc bill run. We do this by updating `ViewBillRunService` to pass `ruleset ` to `ViewBillRunPresenter`. In turn, this passes it to `ViewBillRunInvoicePresenter` (note that presenters can only accept a single `data` parameter so we need to add this to the data we pass in). Finally, this presenter excludes it from the output if the ruleset isn't `presroc`.

* Refactor conditionality

After putting the basic functionality and tests in place, we implement a nicer way of writing the conditionality that allows us to specify it within the response object rather than defining the object and then deleting `minimumChargeInvoice` based on the ruleset.

* Add v3 route