View SRoC Invoice and v3 route (#629)
https://eaflood.atlassian.net/jira/software/projects/CMEA/boards/907?selectedIssue=CMEA-216

We create a GET /v3/{regimeSlug}/bill-runs/{billRunId}/invoices/{invoiceId} route and point it to the same ViewInvoiceService as the v2 route. We don't touch the v2 route as the changes implemented here are not considered to be breaking changes, so we can leave v2 as-is. Then we update the service and its associated presenters to handle SRoC invoices:

ruleset is added to the invoice-level view
the invoice-level minimumChargeInvoice flag is hidden for SRoC invoices
the transaction-level subjectToMinimumCharge & minimumChargeAdjustment flags are also hidden when viewing an SRoC invoice
Applying the deminimis limit (as referred to in the ACs) is handled when the bill run summary is generated, so no additional work is required here.


* View SRoC Invoice and v3 route
https://eaflood.atlassian.net/jira/software/projects/CMEA/boards/907?selectedIssue=CMEA-216

* Set up v3 route

We intend to update the underlying code to support sroc invoices, so we create a v3 route and simply point it to the existing controller

* Add `ruleset` to invoice-level view

We update the query within `ViewInvoiceService` to return the bill run's ruleset and update `ViewInvoicePresenter` to include it in the response

* Make `minimumChargeInvoice` conditional on ruleset

We update `ViewInvoicePresenter` to only return `minimumChargeInvoice` for presroc invoices

* Make `subjectToMinimumCharge` conditional on ruleset

We update `ViewInvoicePresenter` and `ViewLicencePresenter` to pass `ruleset` down to the transaction level, then update `ViewTransactionPresenter` to only return `subjectToMinimumCharge` for presroc invoices

* Make `minimumChargeAdjustment` conditional on ruleset