Add Swagger request versioning to docs (#631)
Recent work updating the CM to support SRoC alongside pre-SRoC has meant that an endpoint can now accept different sets of request parameters depending on which ruleset the request is for. Getting the Swagger syntax right for describing the request formats and providing examples was a little fiddly to implement (not helped by the fact the VS Code extension we use doesn't support previewing multiple sets of examples!) so we document it here for future reference.